requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.5
oauthlib==3.2.2
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.1.0
